% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{load_model}
\alias{load_model}
\title{Load the pre-trained NBA parlay model}
\usage{
load_model()
}
\value{
A `glm` object trained on Elo + AllTimeElo
}
\description{
Load the pre-trained NBA parlay model
}
